+++ front/src/components/layout/index.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import Header from "./Header";
 import b from "components/lib/b";
 
 export default (Component) => class extends React.Component {
@@ -11,19 +12,11 @@ export default (Component) => class extends React.Component {
     render () {
         const {params} = this.props;
         const layout = b.with("mainLayout");
-        const header = b.with("header");
 
         return (
             <div className={layout()}>
                 <div className={layout("headerArea")}>
-                    <header className={header()}>
-                        <h1 className={header("title")}>
-                            <a href="/">{"fumi*fumi"}</a>
-                        </h1>
-                        <div className={header("action")}>
-                            <a href="/magazines/dashboard">{"📈"}</a>
-                        </div>
-                    </header>
+                    <Header />
                 </div>
                 <div className={layout("bodyArea")}>
                     <Component params={params} />
